Yash Patel

+++ RPG6/src/towns/Card.java
@@ -5,27 +5,29 @@ package towns;
  *
  */
 public class Card {
-	private short rank, suit;
+	private int rank, suit;
 
 	    private static String[] suits = { "hearts", "spades", "diamonds", "clubs" };
 	    private static String[] ranks  = { "Ace", "2", "3", "4", "5", "6", "7", "8", "9", "10", "Jack", "Queen", "King" };
 	    
-	    public static String rankAsString( int __rank ) {
+	    public static String rankAsString(int __rank) {
 	        return ranks[__rank];
 	    }
-	    Card(short suit, short rank)
-	    {
+	    
+	    Card(int suit, int rank) {
 	        this.rank=rank;
 	        this.suit=suit;
 	    }
-	    public @Override String toString()
-	    {
+	    
+	    public @Override String toString() {
 	          return ranks[rank] + " of " + suits[suit];
 	    }
-	    public short getRank() {
+	    
+	    public int getRank() {
 	         return rank;
 	    }
-	    public short getSuit() {
+	    public int getSuit() {
 	        return suit;
 	    }
+	    
 	}
